clip4clip/pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["LeeKunHa <kunha98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
torch = "1.7.1"
pandas = "2.0.0"
transformers = "4.7.0"
urllib3 = "1.26.18"
tqdm = "4.66.2"
torchvision = "0.8.2"
pillow = "10.2.0"
requests = "2.31.0"
matplotlib = "3.7.0"
pyparsing = "3.1.2"
glob2 = "^0.7"
natsort = "^8.4.0"
numpy = "1.24.0"
scenedetect = {extras = ["opencv"], version = "^0.6.2"}
svg-path = "^6.3"
pytube = "^15.0.0"
ftfy = "^6.1.3"
pymysql = "*"
opencv-python = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
